requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
collection==0.1.6
discord==2.1.0
discord.py==2.1.0
emoji==2.2.0
frozenlist==1.3.1
gcloud==0.17.0
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.4
jws==0.1.3
multidict==6.0.2
oauth2client==3.0.0
outcome==1.2.0
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.7.0
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.4.0
undetected-chromedriver==3.1.5.post4
urllib3==1.26.12
websockets==10.3
wsproto==1.2.0
yarl==1.8.1
